pyproject.toml
[project]
name = "pwt.dynamic_watch_expression"
description = "此项目通过插件获取数据, 并评估表达式以决定是否执行后续操作."
license = { file = "LICENSE" }
# version = "0.1.0"
dynamic = ["version"]
requires-python = ">= 3.10"
dependencies = [
    "requests",
    "dnspython",
    "pydantic",
    "lark",
    "pwt.utils",
]

# [project.scripts]
# dynwatchexpr = "pwt.dynamic_watch_expression.main:main"

[tool.setuptools.dynamic]
version = { attr = "pwt.dynamic_watch_expression.__version__" }
